Apply changes from review
